factorize formatCodeList
