+++ packages/codelists/src/utils/index.js
@@ -104,3 +104,54 @@ export const recalculatePositions = (codes, tree) => {
 			})),
 	}));
 };
+
+const superiorToParent = (child, parent) => {
+	return child > parent;
+};
+
+const getPosition = (codes, parentName, childPositions) => {
+	const parentPositions = codes.map(
+		([key, value]) => key === parentName && value.positions
+	);
+	return {
+		parent: parentName,
+		position: Math.min(
+			...childPositions.filter((childPos) =>
+				superiorToParent(childPos, parentPositions[0])
+			)
+		),
+	};
+};
+
+export const formatCodeList = (cl) => {
+	if (cl.codes) {
+		cl.codes = Object.values(cl.codes)
+			.sort((a, b) => (a.code > b.code ? 1 : -1))
+			.reduce((acc, c, i) => {
+				return {
+					...acc,
+					[c.code]: {
+						...c,
+						id: c.code,
+						parents:
+							c.parents && c.parents[0]
+								? c.parents.map((p) =>
+										getPosition(
+											Object.entries(cl.codes),
+											p,
+											/* Object.values(Object.values(c.code).positions) */
+											c.positions || [i + 1]
+										)
+								  )
+								: [
+										{
+											parent: '',
+											position: c.positions ? c.positions[0] : i + 1,
+										},
+								  ],
+					},
+				};
+			}, {});
+	}
+	return cl;
+};
